refactor: repository 구현
