+++ src/main/kotlin/jihwan/practice/search/service/KeywordCollectService.kt
@@ -1,22 +1,17 @@
 package jihwan.practice.search.service
 
-import org.springframework.data.redis.core.RedisTemplate
+import jihwan.practice.search.repository.KeywordRepository
 import org.springframework.stereotype.Service
 
 @Service
 class KeywordCollectService(
-    private val redisTemplate: RedisTemplate<String, String>
+    private val keywordRepository: KeywordRepository,
 ) {
-    companion object {
-        private const val KEYWORD_ZSET_KEY = "keyword_counts"
-    }
-
     fun incrementCount(keyword: String) {
-        redisTemplate.opsForZSet().incrementScore(KEYWORD_ZSET_KEY, keyword, 1.0)
+        keywordRepository.incrementCount(keyword)
     }
 
     fun getTopKeywords(size: Int = 10): List<Pair<String, Int>> {
-        val results = redisTemplate.opsForZSet().reverseRangeWithScores(KEYWORD_ZSET_KEY, 0, size - 1.toLong())
-        return results?.map { it.value.toString() to it.score!!.toInt() } ?: emptyList()
+        return keywordRepository.getTopKeywords(size)
     }
 }
